power: qcom: msm-core: Rework resampling processing
This patch reworks the whole resampling logic in the
msm-core driver to rely on a dedicated high priority
workqueue and a timer to handle resampling requests
replacing the inefficient kernel completion event
structure.

Resampling requests which come from the suspend handler
need to be processed unconditionally while the requests
coming from the core_temp notifier need to be validated
by a forced timeout which happens after every resampling
process for poll_ms.

This fixes the issues of constant uninterruptiple waits
coming from the kthread which caused descheduling of other
tasks.

Also the accuracy of the temperature data provided to the
scheduler is improved since we make a direct decision on
whether to trigger a resampling process or not once new
temperature data arrives.

Change-Id: Ib69859d4e17a9528939d9acf7ee6a9fdbda0a0b3
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>