power: qcom: msm-core: Add back periodic polling and fix various race conditions
This adds back the periodic and takes measures to prevent
race conditions such as disabling the forced timeout logic
when direct resampling gets scheduled or while resampling
is in progress.

Additionally we tune the forced timeout to be poll_ms / 2
as we have two possible sources of resampling execution now.

Change-Id: Ie4cabd1bec7e9103843a205c8f4a37f04c4f2a9c
Signed-off-by: Alex Naidis <alex.naidis@paranoidandroid.co>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>