signal: Don't silently convert SI_USER signals to non-current pidfd
The current sys_pidfd_send_signal() silently turns signals with explicit
SI_USER context that are sent to non-current tasks into signals with
kernel-generated siginfo.
This is unlike do_rt_sigqueueinfo(), which returns -EPERM in this case.
If a user actually wants to send a signal with kernel-provided siginfo,
they can do that with pidfd_send_signal(pidfd, sig, NULL, 0); so allowing
this case is unnecessary.

Instead of silently replacing the siginfo, just bail out with an error;
this is consistent with other interfaces and avoids special-casing behavior
based on security checks.

Bug: 135608568
Fixes: 50245d36a814 ("signal: add pidfd_send_signal() syscall")
Test: test program using syscall(__NR_pidfd_send_signal,..) to send SIGKILL
Signed-off-by: Jann Horn <jannh@google.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>