cpufreq: qcom: Limit the rate of CPU clock-rate changes
Without any form of rate limiting on CPU clock-rate changes, the CPU
clock can be changed as frequently as every 50 usecs, which is not
nearly enough time for the hardware to settle into the new rate. This
can result in the CPU performing sub-optimally when the clock-rate is
changed extremely frequently.

To mitigate this, rate-limit CPU clock-rate updates to once every 10 ms.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>