sound: msm: q6adm: Redirect default and custom topologies to NULL topology
Instead of redirecting to default COPP which is equivalent to stereo
speaker COPP topology, let's just nullify it to devoid low latency
playback of any post-DSP processing unless a specific ADM topology is
intentionally set by the routing driver that is not the default nor an
external effect.

Let us also include legacy PCM playback to this patch to completely disable
post-DSP processing as a whole.

Signed-off-by: ederekun <sedrickvince@gmail.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>