wireguard: Revert "compat: skb_mark_not_on_list will be backported to Ubuntu 18.04"
Because this commit has not been backported so far, due to the implications
of building Ubuntu's backport of wireguard in a timely manner.

For now, reverting this fix would allow wireguard-linux-compat CI to work
on Ubuntu 18.04.

A different fix or the same one can be applied again when the time is
right.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>