kbuild: hide commands to run Kconfig, and show short log for silentoldconfig
Some targets (localyesconfig, localmodconfig, defconfig) hide the
command running, but the others do not.

Users know which Kconfig flavor they are running, so it is OK to hide
the command. Add $(Q) to all commands consistently. If you want to see
the full command running, pass V=1 from the command line.

silentoldconfig is the exceptional case, which occurs without explicit
command invocation by the user. Display the Kbuild-style log for it.
The ugly bare log will go away.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>