cpufreq: fallback to schedutil if governor is not found
Nowadays, custom ROMs are starting to set custom governor profiles via
init scripts, often times involving custom governors. If the user sticks
with the built-in kernel, that is totally fine as everything will get
loaded up properly. However, if the user decides to switch to a custom
kernel and those governors are not present, this function will fail to
find them when and the governor will not change. I attempt to work
around this by modifying how the method handles finding the governor.

Process without this patch:
1. Phone boots with other governor.
2. Init scripts attempts to write custom governor to sys path.
3. The function cpufreq_parse_governor fails to locate the
   governor and returns -EINVAL since the input was invalid.

A couple of notes:
1. If you have a different governor that you want to be set on
   boot, make sure to change schedutil to that governor's
   name.
2. This is a little bit of a hack; ideally, it'd be nice to
   only switch to schedutil.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>