ARM: dts: Use haptic node from qcom,qpnp-haptics
Meanwhile, the latest CAF changes use the node "qcom,qpnp-haptics"
which comes from the drivers/leds/leds-qpnp-haptics.c file and it is
enabled by default on sdm660_defconfig file.

I am assuming, if this is not changed it is possible that the
haptic driver will not work or cannot be controlled.

It also adapts to Import from Asus changes.

Signed-off-by: wimbiyoas <wyoashizkia@gmail.com>