nt36xxx: Allocate alloc_workqueue instead create_workqueue()
create_workqueue() is actually an old way of creating workqueues (linux_version < 2.6.36). Patches are already pushed to LWN, suggesting to use alloc_workqueues() instead.
Because the new way of allocating wq instead of creating them is intended to do more than replace workqueues with a cleverer implementation; it is really meant to be a general-purpose task management system for the kernel.

WQ_UNBOUND: workqueues were designed to run tasks on the CPU where they were submitted in the hope that better memory cache behavior would result. This flag turns off that behavior, allowing submitted tasks to be run on any CPU in the system. It is intended for situations where the tasks can run for a long time, to the point that it's better to let the scheduler manage their location. Currently the only user is the object processing code in the FS-Cache subsystem.

WQ_HIGHPRI: tasks submitted to this workqueue will put at the head of the queue and run (almost) immediately. Unlike ordinary tasks, high-priority tasks do not wait for the CPU to become available; they will be run right away. That means that multiple tasks submitted to a high-priority queue may contend with each other for the processor.

Test: Device boots, touchscreen works perfectly

Signed-off-by: Panchajanya1999 <rsk52959@gmail.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>