sched: tune: Track tasks individually to be more accurate
Allocate a boolean in task_struct for schedtune to test in order
to assist in tracking boosted tasks. This is to prevent acquiring
locks unnecessary specially when the specified task was not even
enqueued nor boosted.

Also patch schedtune_can_attach to reflect the boost dependent
de/enqueueing and use the boolean to streamline calls to boost bias.

Signed-off-by: ederekun <sedrickvince@gmail.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>