sched/fair: Do not boost cpu/task utilization if boost = 2
For the same reasons as 6ebac80f8d03a047a46944f26eb2eab4c6e08924 ("sched/fair: Add feature to toggle SchedTune utilization boosting"),
but still allowing a user configuration if a user does want the extra performance the corresponding
power costs.

Choose a boost value of 2 for this as it's not typically set as much as
something like 1, 5, 10, etc. would be used.

Signed-off-by: idkwhoiam322 <idkwhoiam322@raphielgang.org>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>