kernel: Boost to the max for a short amount of time when zygote forks
In order to boost during the period from when a new app is forked to the
time that it is added to the top-app cpuset (which is a signal to
subsequently boost again), boost for a short amount of time when zygote
forks. Not all processes forked from zygote are top apps, hence the
decoupled boosting setup.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>