input/touchscreen: nt36xxx: Place worker thread on SCHED_RR
Since a lot of display related tasks are multiplexed on mdss on a
dedicated thread, they are placed in a FIFO manner, which is a bit
of time consuming since not all those queued tasks have a known or
small execution period. Switching to Round-Robin fashion eliminates
this complexity and bounds on mdss.

Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>