thermal: Add a voltage regulator cooling device
Certain regulators are required to be restricted to a voltage floor on
thermal condition. In order to support this, add a new regulator cooling
device.

This cooling device can place a voltage floor on rpmph ARC type regulator.
Request to place a voltage cap will be ignored by this cooling device.

Change-Id: I0f7128ebadfa83eb7ce8014aa5f1602b04f36672
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>