thermal: core: Use wait_event for sysfs_notify
Change to use wait_event_interruptible instead of wait_for_completion so
that the kthread can be awoken by the kthread_stop call. To properly
handle the kthread_stop, we need to also add it as a condition to the
wait_event parameter.

The use case happens when we try to unregister a power supply device and
we get stuck waiting for kthread_stop trying to kill the sysfs_notify
thread who wouldn't be interrupted until it received a completion.

Once we do break out of the loop; ensure we only send the uevent when we
were actually completed and not because of a kthread stop.

Change-Id: I3e5117cb7c666924c59860deb0f90c531af3bafc
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/792939
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
CRs-fixed: (CR)
Reviewed-on: https://gerrit.mot.com/938674
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Christopher Fries <cfries@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>