qpnp-fg-gen3: Expose property online status
When device enters suspend mode in charging state,
fg_psy_get_property tries to get the online status
but fails to acquire it because of missing the required
property.

dmesg notifies us with the following log:
[1.640123] FG: fg_psy_get_property: unsupported property 4

Add the missing property 4 with a case to define the online status of the driver.

Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>