binder: Use cred instead of task for selinux checks
Since binder was integrated with selinux, it has passed
'struct task_struct' associated with the binder_proc
to represent the source and target of transactions.
The conversion of task to SID was then done in the hook
implementations. It turns out that there are race conditions
which can result in an incorrect security context being used.

Fix by using the 'struct cred' saved during binder_open and pass
it to the selinux subsystem.

Cc: stable@vger.kernel.org # 5.14 (need backport for earlier stables)
Fixes: 79af7307

 ("Add security hooks to binder and implement the hooks for SELinux.")
Suggested-by: default avatarJann Horn <jannh@google.com>
Signed-off-by: default avatarTodd Kjos <tkjos@google.com>
Acked-by: default avatarCasey Schaufler <casey@schaufler-ca.com>
Signed-off-by: default avatarPaul Moore <paul@paul-moore.com>
Change-Id: Id7157515d2b08f11683aeb8ad9b8f1da075d34e7
Bug: 200688826
Signed-off-by: default avatarTodd Kjos <tkjos@google.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>