android: Introduce hwui monitor
This is a pure kernel-based solution to obtain the frame generation
time of Android apps. The frame generation time is a critical metric
for achieving features such as frame-time-aware scheduling, and this
approach allows us to monitor userspace performance based on it. It's
important to note that this measures the actual frame generation time,
not the time between two consecutive frames.

This is achieved by making use of uprobes, which adds breakpoints to
the widely-used libhwui.so library. This means applications with their
own render engine can't be monitored, which involve games and browsers.

Since libhwui.so can differ between different versions and OEMs, it may
be necessary to manually disassemble the library and find the correct
offsets for injections.

The reported ui_frame_time is the time taken by the UI thread before
handing over to the render engine. This means that the time taken on
the render engine and GPU is not included in this measurement. However,
since the most time-consuming part usually occurs on the Java side, this
is not a big concern. Just remember, when choosing the jank_frame_time,
leave a little room for the GPU and render engine.

Signed-off-by: LibXZR <i@xzr.moe>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>