ARM: dts: sdm660: Select efficient freq based on result freqbench
Based on description of the commit [1], we can determine efficient
frequency manual based on the value of the unit I/mJ. In determining
efficient frequency we must choose the highest value from I/mJ unit, if
I/mJ unit value is smaller then the frequency must be skipped or deleted
in order to get an efficient frequency.

But for the lowest freq we won't remove it, as it drains the battery
while the CPU is idle.

Here is the original freqbench result:

===== CPU 1 =====
Frequencies: 633 902 1113 1401 1536 1747 1843

 633:  2058     3.2 C/MHz     48 mW    5.9 J   42.6 I/mJ   121.5 s
 902:  2930     3.2 C/MHz     72 mW    6.2 J   40.6 I/mJ    85.3 s
1113:  3616     3.2 C/MHz     79 mW    5.4 J   46.0 I/mJ    69.1 s
1401:  4551     3.2 C/MHz    125 mW    6.9 J   36.3 I/mJ    54.9 s
1536:  4988     3.2 C/MHz    134 mW    6.7 J   37.1 I/mJ    50.1 s
1747:  5674     3.2 C/MHz    179 mW    7.9 J   31.7 I/mJ    44.1 s

===== CPU 4 =====
Frequencies: 1113 1401 1747 1958 2150 2208

1113:  5825     5.2 C/MHz    220 mW    9.4 J   26.5 I/mJ    42.9 s
1401:  7324     5.2 C/MHz    317 mW   10.8 J   23.1 I/mJ    34.1 s
1747:  9135     5.2 C/MHz    474 mW   13.0 J   19.2 I/mJ    27.4 s
1958: 10247     5.2 C/MHz    578 mW   14.1 J   17.7 I/mJ    24.4 s
2150: 11246     5.2 C/MHz    694 mW   15.4 J   16.2 I/mJ    22.2 s
2208: 11551     5.2 C/MHz    736 mW   15.9 J   15.7 I/mJ    21.7 s

Here are the freqbench results if we use efficient frequencies:

===== CPU 1 =====
Frequencies: 633 1113 1536 1747

 633:  2058     3.2 C/MHz     48 mW    5.9 J   42.6 I/mJ   121.5 s
1113:  3616     3.2 C/MHz     79 mW    5.4 J   46.0 I/mJ    69.1 s
1536:  4988     3.2 C/MHz    134 mW    6.7 J   37.1 I/mJ    50.1 s
1747:  5674     3.2 C/MHz    179 mW    7.9 J   31.7 I/mJ    44.1 s

===== CPU 4 =====
Frequencies: 1113 1747 2208

1113:  5825     5.2 C/MHz    220 mW    9.4 J   26.5 I/mJ    42.9 s
1747:  9135     5.2 C/MHz    474 mW   13.0 J   19.2 I/mJ    27.4 s
2208: 11551     5.2 C/MHz    736 mW   15.9 J   15.7 I/mJ    21.7 s

* [1]: KudProject/kernel_xiaomi_sdm710@e586888
* Source of freqbench result: kdrag0n/freqbench@d59d971

Also adapt for simple-thermal, this is necessary because some freq
has been removed. Fixes same frequency in two different nodes, and
reduces max temperature threshold a little.

Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>