sched: Refactor and rename dynamic_stune to adaptive_tune
Redo the entire structure of the code to improve readability as well as
functionality, remove unnecessary injected code in other drivers.

Implement framebuffer notifier to prevent extending during suspend and
save power as there is no need for reduced system latency.

Introduce PM QoS to impose a low latency requirement to perf power cluster
during active periods to prevent shallow idling and top-app along with
foreground tasks from being slowed down as they get migrated to a perf cluster
CPU, greatly enhancing the latency reduction that task placement manipulation
through schedtune already does to system responsiveness.

Due to these recent changes, the driver is no longer limited to schedtune
adjustments and is more of a general performance tuning driver. Furthermore,
the function of the driver has mostly been more about adapting to user's usage
as much as possible and not only dynamically changing based on an event, which
is why adaptive is more appropriate as a name modifier to our tuning driver.

Of course, lastly, dynamic schedtune is a name already taken by joshuous and I
personally do not want the name to get a bad rep just in case I have done any
sacrilegious addition or change to the driver. I give my utmost respect and
admiration to him as well as the developers that have developed for this device.

Signed-off-by: ederekun <sedrickvince@gmail.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>